replace int by size_t when referencing buffer size in a few places
darktable has various issues when working with huge image sizes. one of the problems
is fixed now but others still remain.

+++ src/develop/pixelpipe_hb.c
@@ -98,7 +98,7 @@ int dt_dev_pixelpipe_init(dt_dev_pixelpipe_t *pipe)
   return res;
 }
 
-int dt_dev_pixelpipe_init_cached(dt_dev_pixelpipe_t *pipe, int32_t size, int32_t entries)
+int dt_dev_pixelpipe_init_cached(dt_dev_pixelpipe_t *pipe, size_t size, int32_t entries)
 {
   pipe->devid = -1;
   pipe->changed = DT_DEV_PIPE_UNCHANGED;
